+++ src/Components/Concerns/HasTimeOptions.php
@@ -8,56 +8,35 @@
  * that is bundled with this source code in the LICENSE file.
  * For details see <https://www.gnu.org/licenses/lgpl-3.0.html>
  */
-
 namespace Noraitec\FilamentFlatpickrPlugin\Components\Concerns;
 
 trait HasTimeOptions
 {
-    public function enableTime(bool $enabled = true): static
+    // Set the time for the calendar
+    public function setTime(string $time): static
     {
-        $this->options['enableTime'] = $enabled;
+        $this->options['time'] = $time;
         return $this;
     }
 
-    public function enableSeconds(bool $enabled = true): static
+    // Set hours for the calendar
+    public function setHours(int $hours): static
     {
-        $this->options['enableSeconds'] = $enabled;
+        $this->options['hours'] = $hours;
         return $this;
     }
 
-    public function time24hr(bool $enabled = true): static
+    // Set minutes for the calendar
+    public function setMinutes(int $minutes): static
     {
-        $this->options['time_24hr'] = $enabled;
+        $this->options['minutes'] = $minutes;
         return $this;
     }
 
-    public function defaultHour(int $hour): static
+    // Clear the time
+    public function clearTime(): static
     {
-        $this->options['defaultHour'] = $hour;
-        return $this;
-    }
-
-    public function defaultMinute(int $minute): static
-    {
-        $this->options['defaultMinute'] = $minute;
-        return $this;
-    }
-
-    public function defaultSeconds(int $seconds): static
-    {
-        $this->options['defaultSeconds'] = $seconds;
-        return $this;
-    }
-
-    public function hourIncrement(int $value): static
-    {
-        $this->options['hourIncrement'] = $value;
-        return $this;
-    }
-
-    public function minuteIncrement(int $value): static
-    {
-        $this->options['minuteIncrement'] = $value;
+        $this->options['time'] = null;
         return $this;
     }
 }